FLEX-1901 Solves small bugs and adds pushtoNewDevices functionality

+++ osgp-adapter-ws-shared-db/src/main/java/com/alliander/osgp/adapter/ws/shared/db/domain/repositories/writable/WritableDeviceModelFirmwareRepository.java
@@ -18,5 +18,5 @@ import com.alliander.osgp.domain.core.entities.DeviceModelFirmware;
 @Repository
 public interface WritableDeviceModelFirmwareRepository extends JpaRepository<DeviceModelFirmware, Long> {
     List<DeviceModelFirmware> findByDeviceModel(DeviceModel deviceModel);
-    DeviceModelFirmware findById(Integer id);
+    DeviceModelFirmware findById(Long id);
 }
